requirements.txt
aiokafka==0.10.0
beanie==1.24.0
fastapi==0.109.0
fastapi-users==12.1.3
fastapi-users-db-beanie==3.0.0
faststream==0.4.3
httpx-oauth==0.13.2
loguru==0.7.2
motor==3.3.2
python-dotenv==1.0.0
uvicorn==0.26.0
